Админка: продукты

+++ server/database/Products.sql
@@ -10,8 +10,8 @@ CREATE TABLE products
 	id SERIAL PRIMARY KEY,
 	barcode TEXT NOT NULL,
 	name TEXT NOT NULL,
-	price NUMERIC NOT NULL,
-	discount NUMERIC NOT NULL,
+	price REAL NOT NULL,
+	discount REAL NOT NULL,
 	amount_type_id INTEGER REFERENCES amount_types(id)
     ON DELETE CASCADE
     ON UPDATE CASCADE
